refactor(package-export-settings): implement new OAS extensions autocomplete component
