+++ packages/portal/src/routes/root/PortalPage/PackageSettingsPage/ExportSettingsTab/ExportSettingsTab.tsx
@@ -2,7 +2,7 @@ import { Box } from '@mui/material'
 import type { FC } from 'react'
 import React, { memo, useCallback } from 'react'
 import type { PackageSettingsTabProps } from '../package-settings'
-import { EditPreservedOasExtensionsDialog } from './EditPreservedOasExtensionsDialog'
+import { EditPreservedOasExtensionsDialog } from './EditPreservedOasExtensionsDialog/EditPreservedOasExtensionsDialog'
 import { useEventBus } from '@apihub/routes/EventBusProvider'
 import { BodyCard } from '@netcracker/qubership-apihub-ui-shared/components/BodyCard'
 import { SettingsEditableParameter } from '../SettingsEditableParameter/SettingsEditableParameter'
